Fix PHPStan level 5 errors (batch 2)
- Console commands: Cast option values to int for subHours/plural calls
- Filesystem: Add ignores for FtpAdapter/SftpAdapter interface quirk
- Foundation: Fix str_replace type, array_filter callback, add type assertion
- Horizon: Change Exception to Throwable for job failures (matches Laravel)
- Horizon: Cast Redis timestamps to string, add higher-order proxy ignore
- Prompts: Add ignores for intentional array_values on lists
- Validation: Add ignores for defensive array_values/array_filter
- Cors: Add ignore for bug (to be fixed in separate PR)

+++ src/Jobs/RetryFailedJob.php
@@ -64,7 +64,7 @@ class RetryFailedJob
         $pushedAt = $payload['pushedAt'] ?? microtime(true);
 
         return $retryUntil
-                        ? CarbonImmutable::now()->addSeconds(ceil($retryUntil - $pushedAt))->getTimestamp()
+                        ? CarbonImmutable::now()->addSeconds((int) ceil($retryUntil - $pushedAt))->getTimestamp()
                         : null;
     }
 }
